requirements.txt
django-currentuser>=0.5.0
django-cors-headers>=3.2.1
djangorestframework>=3.11.0
django-crispy-forms>=1.9.0
whitenoise>=5.0.1